a6010: Add VNDK APEX path as permitted paths of sphal
With VNDK APEX, the path for VNDK libs has been changed
from /system/lib/vndk-VER to /apex/com.android.vndk.vVER/lib

But we need to keep older location since it should run on
older version of system due to APEX requirements.

When VNDK APEX is used, older location is just a symlink to a new
location. So we have to add a new location as permitted path. Otherwise,
linker would fail because it checks accessibility using realpath().

Test: add a vendor binary to /apex/com.android.media.swcodec/bin
      which opens /system/lib/vndk-sp{VNDK_VER}/hw/android.hidl.memory@1.0-impl.so
      via android_load_sphal_library()
Bug: 142912195

Change-Id: I13779679af5cf7f687a726afe0b6118e15cf1869